Fixing issues: #1, #3, #4, #5
- import core.module to app.module\n- moved product/statistics dirs to feature dir\n- moved about.comp to core dir\n- provided services in specific modules
