--- src/app/about/about.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import {AboutRoutingModule} from "./about-routing.module";
-
-@NgModule({
-  declarations: [],
-  imports: [ AboutRoutingModule ],
-  providers: [],
-  bootstrap: []
-})
-export class AboutModule { }
