+++ src/app/app-routing.module.ts
@@ -2,12 +2,12 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 const appRoutes: Routes = [
-  { path: "product", loadChildren: () => import('./product/product.module')
+  { path: "product", loadChildren: () => import('./features/product/product.module')
       .then((x) => x.ProductModule) },
-  { path: "about", loadChildren: () => import('./about/about.module')
-      .then((x) => x.AboutModule) },
-  { path: "statistics", loadChildren: () => import('./statistics/statistics.module')
-      .then((x) => x.StatisticsModule) }
+  { path: "statistics", loadChildren: () => import('./features/statistics/statistics.module')
+      .then((x) => x.StatisticsModule) },
+  { path: 'about', loadComponent: () => import('./core/components/about/about-view.component')
+      .then((x) => x.AboutViewComponent)  }
 ];
 
 @NgModule({
