+++ src/app/app.component.ts
@@ -4,7 +4,6 @@ import { Component } from '@angular/core';
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
-  // dependencies(imports) are in app.module
 })
 export class AppComponent {
   title = 'PharmacyFront';
