+++ src/app/app.module.ts
@@ -4,10 +4,8 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppComponent } from './app.component';
 import {AppRoutingModule} from "./app-routing.module";
 import {provideHttpClient} from "@angular/common/http";
-import { SidenavComponent } from './core/components/sidenav/sidenav.component';
-import { HeaderComponent } from './core/components/header/header.component';
-import { FooterComponent } from './core/components/footer/footer.component';
 import {MatNativeDateModule} from "@angular/material/core";
+import {CoreModule} from "./core/core.module";
 
 @NgModule({
   declarations: [AppComponent],
@@ -15,12 +13,10 @@ import {MatNativeDateModule} from "@angular/material/core";
     BrowserModule,
     BrowserAnimationsModule,
 
+    CoreModule,
     AppRoutingModule,
-    MatNativeDateModule,
 
-    HeaderComponent,
-    SidenavComponent,
-    FooterComponent
+    MatNativeDateModule
   ],
   providers: [provideHttpClient()],
   bootstrap: [AppComponent]
