+++ src/app/core/components/about/about-view.component.html
@@ -1,0 +1,9 @@
+
+<mat-list role="list">
+  <mat-list-item role="listitem">Pharmacy Application</mat-list-item>
+  <mat-list-item role="listitem">Version: 1.0.1</mat-list-item>
+  <mat-list-item role="listitem">Copyrights UrošZarić @ 2024</mat-list-item>
+</mat-list>
+
+
+
