+++ src/app/core/components/footer/footer.component.ts
@@ -1,13 +1,8 @@
 import { Component } from '@angular/core';
-import {MatToolbarModule} from "@angular/material/toolbar";
 
 @Component({
-  standalone: true,
   selector: 'app-footer',
   templateUrl: './footer.component.html',
-  styleUrls: ['./footer.component.css'],
-  imports: [
-    MatToolbarModule
-  ]
+  styleUrls: ['./footer.component.css']
 })
 export class FooterComponent {}
