+++ src/app/core/components/sidenav/sidenav.component.ts
@@ -1,21 +1,9 @@
 import {Component, Input} from '@angular/core';
-import {MatSidenavModule} from "@angular/material/sidenav";
-import {MatListModule} from "@angular/material/list";
-import {MatButtonModule} from "@angular/material/button";
-import {RouterLink, RouterModule} from "@angular/router";
 
 @Component({
-  standalone: true,
   selector: 'app-sidenav',
   templateUrl: './sidenav.component.html',
-  styleUrls: ['./sidenav.component.css'],
-  imports: [
-    RouterModule,
-    RouterLink,
-    MatSidenavModule,
-    MatListModule,
-    MatButtonModule
-  ]
+  styleUrls: ['./sidenav.component.css']
 })
 export class SidenavComponent {
   @Input() messageFromHeader: boolean = false;
