+++ src/app/features/product/components/product-edit/product-edit.component.ts
@@ -3,10 +3,10 @@ import { CommonModule } from "@angular/common";
 import { FormsModule } from "@angular/forms";
 import { Observable, Subscription } from "rxjs";
 import { ActivatedRoute } from "@angular/router";
-import { ProductDialogComponent } from "../../../shared/components/product-dialog/product-dialog.component";
-import { IProduct } from "../../../core/models/iproduct";
-import { IManufacturer } from "../../../core/models/imanufacturer";
-import { ManufacturerService } from "../../../core/services/manufacturer.service";
+import { ProductDialogComponent } from "../../../../shared/components/product-dialog/product-dialog.component";
+import { IProduct } from "../../../../core/models/iproduct";
+import { IManufacturer } from "../../../../core/models/imanufacturer";
+import { ManufacturerService } from "../../../../core/services/manufacturer.service";
 import { ProductService } from "../../services/product.service";
 import { MatSnackBar, MatSnackBarModule } from "@angular/material/snack-bar";
 import { MatFormFieldModule } from "@angular/material/form-field";
