+++ src/app/features/statistics/services/statistic.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { StatisticService } from './statistic.service';
+
+describe('StatisticServiceService', () => {
+  let service: StatisticService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(StatisticService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
