+++ src/app/features/statistics/statistics-routing.module.ts
@@ -1,8 +1,5 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import {ProductViewComponent} from "../product/components/product-view/product-view.component";
-import {AboutViewComponent} from "../about/components/about-view/about-view.component";
-import {StatisticsComponent} from "./components/statistics/statistics.component";
 
 const statisticsRoutes: Routes = [
   { path: '', loadComponent: () => import('./components/statistics/statistics.component')
