--- src/app/product/product.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import {ProductRoutingModule} from "./product-routing.module";
-
-@NgModule({
-  declarations: [],
-  imports: [ ProductRoutingModule ],
-  providers: [],
-  bootstrap: []
-})
-export class ProductModule { }
